feat: report every error that ocures in goroutine to Sentry
