+++ openshift/init_tenant.go
@@ -6,8 +6,9 @@ import (
 	"sync"
 
 	"github.com/fabric8-services/fabric8-common/log"
+	"github.com/fabric8-services/fabric8-tenant/sentry"
 	"github.com/fabric8-services/fabric8-tenant/tenant"
-	yaml "gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v2"
 )
 
 const (
@@ -44,24 +45,21 @@ func RawInitTenant(ctx context.Context, config Config, callback Callback, opensh
 				defer wg.Done()
 				err := ApplyProcessed(Filter(objects, IsOfKind(ValKindProjectRequest, ValKindNamespace)), userOpts)
 				if err != nil {
-					log.Error(ctx, map[string]interface{}{
+					sentry.LogError(ctx, map[string]interface{}{
 						"namespace": namespace,
-						"err":       err,
-					}, "error init user project, ProjectRequest")
+					}, err, "error init user project, ProjectRequest")
 				}
 				err = ApplyProcessed(Filter(objects, IsOfKind(ValKindRoleBindingRestriction)), opts)
 				if err != nil {
-					log.Error(ctx, map[string]interface{}{
+					sentry.LogError(ctx, map[string]interface{}{
 						"namespace": namespace,
-						"err":       err,
-					}, "error init user project, RoleBindingRestrictions")
+					}, err, "error init user project, RoleBindingRestrictions")
 				}
 				err = ApplyProcessed(Filter(objects, IsNotOfKind(ValKindProjectRequest, ValKindNamespace, ValKindRoleBindingRestriction)), userOpts)
 				if err != nil {
-					log.Error(ctx, map[string]interface{}{
+					sentry.LogError(ctx, map[string]interface{}{
 						"namespace": namespace,
-						"err":       err,
-					}, "error init user project, Other")
+					}, err, "error init user project, Other")
 				}
 				_, err = apply(
 					CreateAdminRoleBinding(namespace),
@@ -80,10 +78,9 @@ func RawInitTenant(ctx context.Context, config Config, callback Callback, opensh
 					),
 				)
 				if err != nil {
-					log.Error(ctx, map[string]interface{}{
+					sentry.LogError(ctx, map[string]interface{}{
 						"namespace": namespace,
-						"err":       err,
-					}, "error unable to delete Admin role from project")
+					}, err, "error unable to delete Admin role from project")
 				}
 			}(key, val, masterOpts, userOpts)
 		} else {
@@ -91,10 +88,9 @@ func RawInitTenant(ctx context.Context, config Config, callback Callback, opensh
 				defer wg.Done()
 				err := ApplyProcessed(objects, opts)
 				if err != nil {
-					log.Error(ctx, map[string]interface{}{
+					sentry.LogError(ctx, map[string]interface{}{
 						"namespace": namespace,
-						"err":       err,
-					}, "error dsaas project")
+					}, err, "error dsaas project")
 				}
 			}(key, val, masterOpts)
 		}
@@ -131,11 +127,10 @@ func RawUpdateTenant(ctx context.Context, config Config, callback Callback, user
 				opts.WithNamespace(namespace),
 			)
 			if err != nil {
-				log.Error(ctx, map[string]interface{}{
+				sentry.LogError(ctx, map[string]interface{}{
 					"output":    output,
 					"namespace": namespace,
-					"error":     err,
-				}, "failed")
+				}, err, "ns update failed")
 				return
 			}
 			log.Info(ctx, map[string]interface{}{
